Introduce an `AsyncConnectionWrapper` type
This type turns a `diesel_async::AsyncConnection` into a
`diesel::Conenction`. I see the following use-cases for this:

* Having a pure rust sync diesel connection implementation for postgres
and mysql can simplify the setup of new diesel projects
* Allowing projects depending on `diesel_async` to use
`diesel_migrations` without depending on `libpq`/`libmysqlclient`

This change requires restructuring the implementation of
`AsyncPgConnection` a bit so that we make the returned future `Send`
independently of whether or not the query parameters are `Send`. This is
possible by serialising the bind parameters before actually constructing
the future.

It also refactors the `TransactionManager` implementation to share more
code with diesel itself.