Fix #198
This commit introduces a boolean flag that tracks whether we currently
execute a transaction related SQL command. We set this flag to true
directly before starting the future execution and back to false
afterwards.
This enables us to detect the cancellation of such futures while the
command is executed. In such cases we consider the connection to be
broken as we do not know how much of the command was actually executed.