Fix a bug in how we handle serialization errors
This commit fixes a bug in how we handle serialization errors with the
postgres backend and transactions. It turns out that we never update the
transaction manager state for `batch_execute` calls, which in turn are
used for executing the transaction SQL itself. That could lead to
situations in which we don't roll back the transaction, but in which we should
have done that.

Fixes #241