Replace hand written future impls
This commit replaces the hand written future impls with such ones that
wrap the types from futures-util. This hopefully reduces the complexity
and also allows using the well tested types instead of hand written
ones.

I also removed the dependency on pin-project-light in favour of two
lines of unsafe code as pin-project would do the same thing internally
as well and these two instances are trivial to reason about.