Keep track of mysql statements to close
This change refactors how we close uncached mysql statements. Instead of
trying to issue the close request as part of the same SQL query, we now
just keep track of whether a statement need to be closed or not and
execute the closing as first operation in the next query execution. The
large advantage of this approach is that we sidestep any async
drop/cancelation related problems. The disadvantage is that we keep that
statement open for a bit longer. I cannot see how we ever would have
more than one statement in there per connection and we also keep the
cached statements around, so this shouldn't cause any problems in
practice.

Fix #269